backend/src/database/database.py
from sqlmodel import create_engine, Session
from typing import Generator
from ..models.user import User
from ..models.task import Task
from sqlmodel import SQLModel
import os
from dotenv import load_dotenv

load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL", "postgresql://neondb_owner:npg_KjN5CRXP0ZAl@ep-polished-bonus-aho16gac-pooler.c-3.us-east-1.aws.neon.tech/neondb?sslmode=require&channel_binding=require")

engine = create_engine(DATABASE_URL, echo=True)

def create_db_and_tables():
    SQLModel.metadata.create_all(engine)

def get_session() -> Generator[Session, None, None]:
    with Session(engine) as session:
        yield session